// frontend/src/api/generated.ts
export type Client = {
  alias: string;
  ip_address: string;
  is_member: boolean;
  last_seen: string;
  query_count: number;
};
// ...
export type Group = {
  blocklist_count: number;
  id: number;
  is_member: boolean;
  member_count: number;
  name: string;
};
// ...
export type PolicyView = {
  allowlist_count: number;
  blocklist_count: number;
  description: string;
  id: number;
  name: string;
  usage_count: number;
};
// ...
export type RangeView = { cidr: string; created_at: string; id: number; name: string };